src/app/client/src/app/modules/certificate/components/certificate-configuration/certificate-configuration.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { CertificateConfigurationComponent } from './certificate-configuration.component';

describe('CertificateConfigurationComponent', () => {
  let component: CertificateConfigurationComponent;
  let fixture: ComponentFixture<CertificateConfigurationComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ CertificateConfigurationComponent ]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(CertificateConfigurationComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    component.ngOnInit();
    expect(component).toBeTruthy();
  });

  it('should terminate all subscriptions', () => {
    spyOn(component.unsubscribe$, 'next');
    spyOn(component.unsubscribe$, 'complete');
    component.ngOnDestroy();
    expect(component.unsubscribe$.next).toHaveBeenCalled();
    expect(component.unsubscribe$.complete).toHaveBeenCalled();

  });

  it('should fetch certificate details', () => {
    component.getCertificateDetails();
  });

  it('should add certificate to the batch', () => {
    component.addCertificateToBatch();
  });
});
